refactored about page and moved coretenets to seperate component. Updated lazy loading.
